requirements.txt
openai-whisper
pyaudio
wave